+++ backend/call-service/Cargo.toml
@@ -14,7 +14,6 @@ jsonwebtoken = "9.3.1"
 reqwest = { version = "0.12.22", features = ["json"] }
 serde = {version = "1.0.219", features = ["derive"] }
 serde_json = "1.0.141"
-smol = "2.0.2"
 sqlx = {version = "0.8.6",features = ["postgres", "runtime-tokio"]}
 thiserror = "2.0.12"
 tokio = {version = "1.46.1", features = ["full"]}
