+++ backend/call-service/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod app;
+pub(crate) mod appstate;
 pub(crate) mod jwt;
 pub(crate) mod room;
 pub(crate) mod sql_utils;
