+++ backend/call-service/src/sql_utils/calls.rs
@@ -1,7 +1,11 @@
 use sqlx::PgPool;
+use tokio::sync::mpsc;
 use topic_structs::{
     GroupCreatedEvent, GroupDeletedEvent, GroupUserAddedEvent, GroupUserRemovedEvent,
 };
+use tracing::debug;
+
+use crate::{appstate::AppStateMessage, room::Room};
 
 pub async fn create_group(pool: &PgPool, event: GroupCreatedEvent) {
     sqlx::query(
@@ -18,7 +22,7 @@ pub async fn create_group(pool: &PgPool, event: GroupCreatedEvent) {
     for user_id in event.member_ids {
         let event = GroupUserAddedEvent {
             group_id: event.group_id.clone(),
-            user_id: user_id,
+            user_id,
         };
 
         add_user_to_group(pool, event).await;
@@ -89,4 +93,28 @@ pub async fn remove_user_from_group(pool: &PgPool, event: GroupUserRemovedEvent)
     .execute(pool)
     .await
     .ok();
+}
+
+pub async fn get_room_from_db(
+    db: &PgPool,
+    room_id: &str,
+    sender: mpsc::Sender<AppStateMessage>,
+) -> sqlx::Result<Room> {
+    debug!("Searching for db room: {}", room_id);
+    let users: Vec<String> = sqlx::query_scalar(
+        "
+        SELECT u.id
+        FROM rooms r
+        JOIN in_room ir ON ir.room_id = r.id
+        JOIN users u ON ir.user_id = u.id
+        WHERE r.id = $1
+    ",
+    )
+    .bind(room_id)
+    .fetch_all(db)
+    .await?;
+
+    debug!("Room fetched from database with users: {:?}", users);
+
+    Ok(Room::new(users, sender, room_id.to_string()).await)
 }
