+++ backend/call-service/src/sql_utils/mod.rs
@@ -1,1 +1,2 @@
 pub(crate) mod calls;
+pub(crate) mod init;
